+++ api/server.js
@@ -10,15 +10,19 @@ const config = require('./src/constants/config');
 const shutdown = require('./src/util/shutdown.util');
 const heartbeat = require('./src/util/heartbeat.util');
 const validate = require('./src/schemas/validate');
+const opencv = require('./src/util/opencv');
 
 module.exports.start = async () => {
   config.setup();
   storage.setup();
   console.log(`Double Take v${version}`);
-  validate(config());
   console.verbose(config());
+  validate(config());
   await database.init();
-  const server = http.Server(require('./src/app')).listen(SERVER.PORT);
+  const server = http.Server(require('./src/app')).listen(SERVER.PORT, async () => {
+    console.verbose(`api listening on :${SERVER.PORT}`);
+    if (opencv.shouldLoad()) await opencv.load();
+  });
   mqtt.connect();
   storage.purge();
   socket.connect(server);