+++ api/src/middlewares/index.js
@@ -1,9 +1,7 @@
-const expressValidator = require('express-validator');
+const Joi = require('joi');
 const { auth, jwt } = require('../util/auth.util');
 const { UNAUTHORIZED } = require('../constants/http-status');
 const { AUTH } = require('../constants')();
-
-const { validationResult } = expressValidator;
 
 module.exports.jwt = async (req, res, next) => {
   try {
@@ -32,15 +30,41 @@ module.exports.setup = async (req, res, next) => {
   }
 };
 
-module.exports.validate = (checks) => [
-  ...checks,
-  (req, res, next) => {
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) {
-      return res.status(400).json({ error: 'Unprocessable Entity', errors: errors.array() });
+module.exports.validate = (schemas) => (req, res, next) => {
+  const errors = [];
+  for (const [key, joiSchema] of Object.entries(schemas)) {
+    const { allowUnknown, ...schema } = joiSchema;
+    const isArray = Array.isArray(req[key]);
+    const { error, value } = isArray
+      ? joiSchema.schema.validate([...req[key]], {
+          allowUnknown,
+          abortEarly: false,
+        })
+      : Joi.object(schema).validate(
+          { ...req[key] },
+          {
+            allowUnknown:
+              key === 'query' && allowUnknown === undefined ? true : allowUnknown !== undefined,
+            abortEarly: false,
+          }
+        );
+    if (error?.details) {
+      errors.push(
+        ...error.details.map((obj) => ({
+          location: key,
+          key: obj.context.label,
+          error: obj.message,
+        }))
+      );
     }
-    next();
-  },
-];
 
-module.exports.expressValidator = expressValidator;
+    // update request to use validate values which may be transformed
+    req[key] = value;
+  }
+
+  if (errors.length) return res.status(422).send({ errors });
+
+  next();
+};
+
+module.exports.Joi = Joi;