+++ api/src/routes/auth.routes.js
@@ -1,21 +1,35 @@
 const express = require('express');
-const { jwt, setup, validate, expressValidator } = require('../middlewares');
+const { jwt, setup, validate, Joi } = require('../middlewares');
 const controller = require('../controllers/auth.controller');
-
-const { body } = expressValidator;
 
 const router = express.Router();
 
 router
-  .post('/', validate([body('password').isLength({ min: 1 })]), controller.login)
-  .post('/password', validate([body('password').isLength({ min: 1 })]), setup, controller.password)
+  .post('/', validate({ body: { password: Joi.string().min(1).required() } }), controller.login)
+  .post(
+    '/password',
+    validate({ body: { password: Joi.string().min(1).required() } }),
+    setup,
+    controller.password
+  )
   .patch(
     '/password',
-    validate([body('password').isLength({ min: 1 }), body('newPassword').isLength({ min: 1 })]),
+    jwt,
+    validate({
+      body: {
+        password: Joi.string().min(1).required(),
+        newPassword: Joi.string().min(1).required(),
+      },
+    }),
     controller.updatePassword
   )
   .get('/tokens', jwt, controller.tokens.get)
   .post('/tokens', jwt, controller.tokens.create)
-  .delete('/tokens/:token', jwt, controller.tokens.delete);
+  .delete(
+    '/tokens/:token',
+    jwt,
+    validate({ params: { token: Joi.string().uuid() } }),
+    controller.tokens.delete
+  );
 
 module.exports = router;