+++ api/src/routes/camera.routes.js
@@ -1,19 +1,18 @@
 const express = require('express');
-const { jwt, validate, expressValidator } = require('../middlewares');
+const { jwt, validate, Joi } = require('../middlewares');
 const controller = require('../controllers/camera.controller');
-
-const { param, query } = expressValidator;
 
 const router = express.Router();
 
 router.get(
   '/:camera',
-  validate([
-    param('camera').isLength({ min: 1 }),
-    query('break').default(true).isIn([true, false]),
-    query('attempts').default(1).isInt().withMessage('not a valid number'),
-  ]),
   jwt,
+  validate({
+    query: {
+      break: Joi.string().valid('true', 'false').default(true),
+      attempts: Joi.number().integer().default(1).min(1),
+    },
+  }),
   controller.event
 );