+++ frontend/src/views/Match.vue
@@ -267,7 +267,7 @@ export default {
               position: 'top',
               accept: async () => {
                 try {
-                  await ApiService.delete('match', { data: $this.matches.selected });
+                  await ApiService.delete('match', { data: { ids: $this.matches.selected } });
                   $this.pagination.total -= $this.matches.selected.length;
                   $this.dropdowns.total -= $this.matches.selected.length;
                   const { areAllSelected } = $this;